Drop ENABLE_UPDATED_DAPP_CONNECTIONS feature flag
It is enabled and shall remain so!
